EPMEDU-728 - Merged conflicts
