+++ feature/more/src/main/java/com/epmedu/animeal/more/TabMoreScreen.kt
@@ -8,32 +8,29 @@ import com.epmedu.animeal.more.about.AboutScreen
 import com.epmedu.animeal.more.donate.DonateScreen
 import com.epmedu.animeal.more.help.HelpScreen
 import com.epmedu.animeal.more.profile.ProfileScreen
-import com.google.accompanist.navigation.animation.AnimatedNavHost
-import com.google.accompanist.navigation.animation.composable
-import com.google.accompanist.navigation.animation.rememberAnimatedNavController
+import com.epmedu.animeal.more.root.MoreScreen
+import com.epmedu.animeal.navigation.AnimatedScreenNavHost
+import com.epmedu.animeal.resources.R
 
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
-fun MoreScreen() {
-    val navController = rememberAnimatedNavController()
-
-    AnimatedNavHost(
-        navController = navController,
+fun TabMoreScreen() {
+    AnimatedScreenNavHost(
         startDestination = NavigationScreen.More.route.name,
         enterTransition = { slideIntoContainer(AnimatedContentScope.SlideDirection.Left) },
         exitTransition = { slideOutOfContainer(AnimatedContentScope.SlideDirection.Right) }
     ) {
-        composable(
+        screen(
             NavigationScreen.More.route.name,
             enterTransition = { slideIntoContainer(AnimatedContentScope.SlideDirection.Right) },
             exitTransition = { slideOutOfContainer(AnimatedContentScope.SlideDirection.Left) }
         ) {
-            MoreScreenUi { route -> navController.navigate(route) }
+            MoreScreen()
         }
-        composable(NavigationScreen.ProfilePage.route.name) { ProfileScreen(navController) }
-        composable(NavigationScreen.Donate.route.name) { DonateScreen(navController) }
-        composable(NavigationScreen.Help.route.name) { HelpScreen(navController) }
-        composable(NavigationScreen.About.route.name) { AboutScreen(navController) }
+        screen(NavigationScreen.ProfilePage.route.name) { ProfileScreen() }
+        screen(NavigationScreen.Donate.route.name) { DonateScreen() }
+        screen(NavigationScreen.Help.route.name) { HelpScreen() }
+        screen(NavigationScreen.About.route.name) { AboutScreen() }
     }
 }
 
@@ -46,10 +43,10 @@ internal val screens = listOf(
 
 internal sealed class NavigationScreen(val route: Route, @StringRes val title: Int) {
     object More : NavigationScreen(route = Route.MORE, title = R.string.more)
-    object ProfilePage : NavigationScreen(route = Route.PROFILE, title = R.string.profile)
-    object Donate : NavigationScreen(route = Route.DONATE, title = R.string.donate)
-    object Help : NavigationScreen(route = Route.HELP, title = R.string.help)
-    object About : NavigationScreen(route = Route.ABOUT, title = R.string.about_detailed)
+    object ProfilePage : NavigationScreen(route = Route.PROFILE, title = R.string.page_profile)
+    object Donate : NavigationScreen(route = Route.DONATE, title = R.string.page_donate)
+    object Help : NavigationScreen(route = Route.HELP, title = R.string.page_help)
+    object About : NavigationScreen(route = Route.ABOUT, title = R.string.page_about_detailed)
 
     enum class Route { MORE, PROFILE, DONATE, HELP, ABOUT }
 }
