+++ feature/more/src/main/java/com/epmedu/animeal/more/root/MoreScreen.kt
@@ -1,18 +1,18 @@
-package com.epmedu.animeal.more.profile
+package com.epmedu.animeal.more.root
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.epmedu.animeal.extensions.currentOrThrow
+import com.epmedu.animeal.more.MoreViewModel
 import com.epmedu.animeal.navigation.navigator.LocalNavigator
 
-
 @Composable
-internal fun ProfileScreen() {
-    val viewModel: ProfileViewModel = viewModel()
+internal fun MoreScreen() {
+    val viewModel: MoreViewModel = viewModel()
     val navigator = LocalNavigator.currentOrThrow
 
-    ProfileScreenUI(
-        onBack = navigator::popBackStack,
-        onEdit = viewModel::edit
+    MoreScreenUi(
+        onLogout = viewModel::logout,
+        onNavigate = navigator::navigate,
     )
 }
