+++ feature/more/src/main/java/com/epmedu/animeal/more/root/ui/MoreOption.kt
@@ -1,0 +1,46 @@
+package com.epmedu.animeal.more.root.ui
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Icon
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.epmedu.animeal.base.theme.AnimealTheme
+
+@Composable
+internal fun MoreOption(title: String, onClick: () -> Unit) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(40.dp)
+            .clickable { onClick() }
+            .padding(start = 44.dp, end = 32.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Text(text = title)
+        Icon(
+            imageVector = Icons.Default.KeyboardArrowRight,
+            contentDescription = title,
+            modifier = Modifier.size(32.dp)
+        )
+    }
+}
+
+@Suppress("UnusedPrivateMember")
+@Preview
+@Composable
+private fun MoreOptionPreview() {
+    AnimealTheme {
+        Surface {
+            MoreOption("Profile Page") {}
+        }
+    }
+}
