+++ feature/more/src/main/java/com/epmedu/animeal/more/ui/common/ScreenPlaceholder.kt
@@ -4,8 +4,8 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.epmedu.animeal.base.ui.BackButton
-import com.epmedu.animeal.base.ui.TopBar
+import com.epmedu.animeal.foundation.topbar.BackButton
+import com.epmedu.animeal.foundation.topbar.TopBar
 
 @Composable
 internal fun ScreenPlaceholder(title: String, onBack: () -> Unit) {
