+++ library/foundation/src/main/java/com/epmedu/animeal/foundation/topbar/TopBar.kt
@@ -2,20 +2,6 @@ package com.epmedu.animeal.foundation.topbar
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.KeyboardArrowLeft
 import androidx.compose.runtime.Composable
@@ -70,9 +56,10 @@ fun BackButton(onClick: () -> Unit) {
     }
 }
 
+@Suppress("UnusedPrivateMember")
 @Preview
 @Composable
-fun TopBarPreview() {
+private fun TopBarPreview() {
     AnimealTheme {
         Surface {
             Column {
